extract PersonCards to their own file
